+++ packages/components/src/Form/Inputs/Combobox/ComboboxList.tsx
@@ -53,7 +53,6 @@ import { ComboboxOptionIndicatorProps } from './ComboboxOptionIndicator'
 import { ComboboxContext, ComboboxMultiContext } from './ComboboxContext'
 import { useBlur } from './utils/useBlur'
 import { useKeyDown } from './utils/useKeyDown'
-import { ComboboxActionType } from './utils/state'
 
 export interface ComboboxListProps
   extends Pick<ComboboxOptionIndicatorProps, 'indicator'>,
@@ -86,6 +85,11 @@ export interface ComboboxListProps
    * @default false
    */
   windowedOptions?: boolean
+  /**
+   * Whether to honor the first click outside the popover
+   * @default false
+   */
+  cancelClickOutside?: boolean
 }
 
 interface ComboboxListInternalProps extends ComboboxListProps {
@@ -106,6 +110,9 @@ const ComboboxListInternal = forwardRef(
       closeOnSelect = true,
       // disables the optionsRef behavior, to be handled externally (support keyboard nav in long lists)
       windowedOptions = false,
+      // passed to usePopover – when false, allows first outside click to be honored
+      // generally should be false except for when closely mimicking native browser select
+      cancelClickOutside = false,
       indicator,
       isMulti,
       ...props
@@ -120,7 +127,6 @@ const ComboboxListInternal = forwardRef(
       closeOnSelectPropRef,
       windowedOptionsPropRef,
       indicatorPropRef,
-      transition,
       wrapperElement,
       isVisible,
       optionsRef,
@@ -177,12 +183,14 @@ const ComboboxListInternal = forwardRef(
 
     const setOpen = (isOpen: boolean) => {
       if (!isOpen) {
-        transition && transition(ComboboxActionType.BLUR)
+        // Without passing an event, this just handles state change required when closing the list
+        handleBlur()
       }
     }
 
     const { popover, contentContainer, popperInstanceRef } = usePopover({
       arrow: false,
+      cancelClickOutside,
       content,
       focusTrap: false,
       isOpen: isVisible,